+++ src/interfaces/IJBDirectoryAccessControl.sol
@@ -2,6 +2,6 @@
 pragma solidity ^0.8.0;
 
 interface IJBDirectoryAccessControl {
-    function setTerminalsAllowed(uint256 projectId) external view returns (bool);
-    function setControllerAllowed(uint256 projectId) external view returns (bool);
+    function setTerminalsAllowed(uint32 projectId) external view returns (bool);
+    function setControllerAllowed(uint32 projectId) external view returns (bool);
 }
