+++ src/interfaces/IJBMigratable.sol
@@ -4,5 +4,5 @@ pragma solidity ^0.8.0;
 import {IERC165} from "@openzeppelin/contracts/utils/introspection/IERC165.sol";
 
 interface IJBMigratable is IERC165 {
-    function receiveMigrationFrom(IERC165 from, uint256 projectId) external;
+    function receiveMigrationFrom(IERC165 from, uint32 projectId) external;
 }
