+++ src/libraries/JBRulesetMetadataResolver.sol
@@ -6,18 +6,18 @@ import {JBRulesetMetadata} from "./../structs/JBRulesetMetadata.sol";
 import {JBConstants} from "./JBConstants.sol";
 
 library JBRulesetMetadataResolver {
-    function reservedRate(JBRuleset memory ruleset) internal pure returns (uint256) {
-        return uint256(uint16(ruleset.metadata >> 4));
+    function reservedRate(JBRuleset memory ruleset) internal pure returns (uint16) {
+        return uint16(ruleset.metadata >> 4);
     }
 
-    function redemptionRate(JBRuleset memory ruleset) internal pure returns (uint256) {
+    function redemptionRate(JBRuleset memory ruleset) internal pure returns (uint16) {
         // Redemption rate is a number 0-10000.
-        return uint256(uint16(ruleset.metadata >> 20));
+        return uint16(ruleset.metadata >> 20);
     }
 
-    function baseCurrency(JBRuleset memory ruleset) internal pure returns (uint256) {
+    function baseCurrency(JBRuleset memory ruleset) internal pure returns (uint32) {
         // Currency is a number 0-4294967296.
-        return uint256(uint32(ruleset.metadata >> 36));
+        return uint32(ruleset.metadata >> 36);
     }
 
     function pausePay(JBRuleset memory ruleset) internal pure returns (bool) {
@@ -79,13 +79,13 @@ library JBRulesetMetadataResolver {
         // version 1 in the bits 0-3 (4 bits).
         packed = 1;
         // reserved rate in bits 4-19 (16 bits).
-        packed |= rulesetMetadata.reservedRate << 4;
+        packed |= uint256(rulesetMetadata.reservedRate) << 4;
         // redemption rate in bits 20-35 (16 bits).
         // redemption rate is a number 0-10000.
-        packed |= rulesetMetadata.redemptionRate << 20;
+        packed |= uint256(rulesetMetadata.redemptionRate) << 20;
         // base currency in bits 36-67 (32 bits).
         // base currency is a number 0-16777215.
-        packed |= rulesetMetadata.baseCurrency << 36;
+        packed |= uint256(rulesetMetadata.baseCurrency) << 36;
         // pause pay in bit 68.
         if (rulesetMetadata.pausePay) packed |= 1 << 68;
         // pause credit transfers in bit 69.
