+++ test/TestPayHooks.sol
@@ -6,7 +6,7 @@ import /* {*} from */ "./helpers/TestBaseWorkflow.sol";
 contract TestPayHooks_Local is TestBaseWorkflow {
     uint8 private constant _WEIGHT_DECIMALS = 18;
     uint8 private constant _NATIVE_TOKEN_DECIMALS = 18;
-    uint256 private constant _WEIGHT = 1000 * 10 ** _WEIGHT_DECIMALS;
+    uint88 private constant _WEIGHT = uint88(1000 * 10 ** _WEIGHT_DECIMALS);
     uint256 private constant _DATA_HOOK_WEIGHT = 2000 * 10 ** _WEIGHT_DECIMALS;
     address private constant _DATA_HOOK = address(bytes20(keccak256("datahook")));
     bytes private constant _PAYER_METADATA = bytes("Some payer metadata");
@@ -17,7 +17,7 @@ contract TestPayHooks_Local is TestBaseWorkflow {
     address private _beneficiary;
     address private _payer;
 
-    uint256 _projectId;
+    uint32 _projectId;
 
     function setUp() public override {
         super.setUp();
